requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
anndata=0.9.2=pypi_0
blosc2=2.0.0=pypi_0
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
einops=0.8.0=pypi_0
fairscale=0.4.13=pypi_0
filelock=3.15.4=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.6.1=pypi_0
google-auth=2.34.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.66.1=pypi_0
h5py=3.11.0=pypi_0
huggingface-hub=0.0.8=pypi_0
idna=3.8=pypi_0
importlib-metadata=8.4.0=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
legacy-api-wrap=1.4=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.41.1=pypi_0
local-attention=1.9.14=pypi_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.6.3=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numba=0.58.1=pypi_0
numexpr=2.8.6=pypi_0
numpy=1.22.0=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.68=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.14=h5eee18b_0
packaging=24.1=pypi_0
pandas=1.1.5=pypi_0
patsy=0.5.6=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py38h06a4308_0
protobuf=5.28.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pynndescent=0.5.13=pypi_0
pyparsing=3.1.4=pypi_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.7.24=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
sacremoses=0.1.1=pypi_0
safetensors=0.4.4=pypi_0
scanpy=1.7.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.12.2=pypi_0
setuptools=72.1.0=py38h06a4308_0
sinfo=0.3.4=pypi_0
six=1.16.0=pypi_0
sqlite=3.45.3=h5eee18b_0
statsmodels=0.14.1=pypi_0
stdlib-list=0.10.0=pypi_0
sympy=1.13.2=pypi_0
tables=3.8.0=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=1.0.9=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.10.3=pypi_0
torch=2.4.0=pypi_0
torchaudio=2.4.0=pypi_0
torchvision=0.19.0=pypi_0
tqdm=4.66.5=pypi_0
transformers=4.6.1=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
umap-learn=0.5.6=pypi_0
urllib3=2.2.2=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.20.1=pypi_0
zlib=1.2.13=h5eee18b_1
